Separate lighting objects top control from low
