+++ DestinationDeepSpace/src/main/java/frc/robot/subsystem/lighting/LightingSubsystem.java
@@ -8,7 +8,6 @@
 package frc.robot.subsystem.lighting;
 
 import frc.robot.subsystem.BitBucketSubsystem;
-import frc.robot.subsystem.lighting.LightingControl.LightingObjects;
 
 /**
  * Add your docs here.
@@ -29,7 +28,68 @@ public class LightingSubsystem extends BitBucketSubsystem {
 	{
 		setName("LightingSubsystem");
 	}
+
+	public enum LightingObjects
+	{
+		// Currently planning on lighting on these controls
+		VISION_SUBSYSTEM(0),
+		DRIVE_SUBSYSTEM(1),
+		SCORING_SUBSYSTEM(2),
+		CLIMB_SUBSYSTEM(3);
+		// RESERVED 4 - 9
+		
+		private int value;
+		
+		LightingObjects(int value)
+		{
+			this.value = value;
+		}
+		
+		public int getValue() { return value; }
+	};	
+	
 	private LightingControl lightingControl;
+
+	public void setAllOff()
+	{
+		for (LightingObjects lightingObject: LightingObjects.values())
+		{
+			lightingControl.setOff(lightingObject.getValue());
+		}
+	}
+	public void setAll(String function, String color, int nspace, int period_msec)
+	{
+		for (LightingObjects lightingObject: LightingObjects.values())
+		{
+			lightingControl.set(lightingObject.getValue(),
+								function,
+								color,
+								nspace,
+								period_msec);
+		}
+	}
+	public void setAllSleeping()
+	{
+		setAll(LightingControl.FUNCTION_SNORE,
+			   LightingControl.COLOR_VIOLET,
+			   0,	// nspace - don't care
+			   0);	// period_msec - don't care
+	}
+	public void setAllSparkles(int period_msec)
+	{
+		setAll(LightingControl.FUNCTION_SPARKLES,
+		       LightingControl.COLOR_BLACK,			// don't care, sparkles are random
+			   0,	            	// nspace - don't care
+			   period_msec);		// period_msec - nice
+	}
+	public void setAllSparkles()	// Default period
+	{
+		setAll(LightingControl.FUNCTION_SPARKLES,
+		       LightingControl.COLOR_BLACK,		// don't care, sparkles are random
+			   0,     			// nspace - don't care
+			   100) ; 			// period_msec - nice default
+	}
+
 
   	@Override
 	public void diagnosticsInit() {
@@ -68,7 +128,7 @@ public class LightingSubsystem extends BitBucketSubsystem {
 
 		lightingControl = new LightingControl();
 
-		lightingControl.setAllSleeping();
+		setAllSleeping();
 
 	}
